+++ simba/video_processors/roi_selector_polygon.py
@@ -1,16 +1,17 @@
 import os
 from typing import Optional, Tuple, Union
+from shapely.geometry import Polygon
+from scipy.spatial import ConvexHull
+from scipy.spatial.distance import cdist
 
 import cv2
 import numpy as np
 
-from simba.utils.checks import (check_file_exist_and_readable,
-                                check_if_valid_img, check_if_valid_rgb_tuple,
-                                check_int, check_str)
+from simba.utils.checks import (check_file_exist_and_readable, check_if_valid_img, check_if_valid_rgb_tuple, check_int, check_str)
 from simba.utils.enums import Options
 from simba.utils.errors import InvalidFileTypeError
 from simba.utils.read_write import get_fn_ext, read_frm_of_video
-from simba.utils.warnings import CropWarning
+from simba.utils.warnings import ROIWarning
 
 
 class ROISelectorPolygon(object):
@@ -37,17 +38,17 @@ class ROISelectorPolygon(object):
     >>> polygon_selector.run()
     """
 
-    def __init__(
-        self,
-        path: Union[str, os.PathLike],
-        thickness: int = 2,
-        clr: Tuple[int, int, int] = (147, 20, 255),
-        title: Optional[str] = None,
-    ) -> None:
+    def __init__(self,
+                 path: Union[str, os.PathLike, np.ndarray],
+                 thickness: int = 2,
+                 vertice_size: int = 3,
+                 clr: Tuple[int, int, int] = (147, 20, 255),
+                 title: Optional[str] = None,
+                 destroy: bool = True) -> None:
 
-        check_file_exist_and_readable(file_path=path)
         check_if_valid_rgb_tuple(data=clr)
         check_int(name="Thickness", value=thickness, min_value=1, raise_error=True)
+        check_int(name="vertice_size", value=vertice_size, min_value=1, raise_error=True)
         if title is not None:
             check_str(name="ROI title window", value=title)
 
@@ -75,7 +76,7 @@ class ROISelectorPolygon(object):
         else:
             self.title = title
 
-        self.drawing, self.clr, self.thickness = False, clr, thickness
+        self.drawing, self.clr, self.thickness, self.destroy, self.vertice_size = False, clr, thickness, destroy, vertice_size
         self.polygon_vertices = []
 
     def draw_polygon(self, event, x, y, flags, param):
@@ -86,25 +87,13 @@ class ROISelectorPolygon(object):
             if self.drawing:
                 temp_img = self.image.copy()
                 for i in range(len(self.polygon_vertices) - 1):
-                    cv2.line(
-                        temp_img,
-                        self.polygon_vertices[i],
-                        self.polygon_vertices[i + 1],
-                        self.clr,
-                        self.thickness,
-                    )
+                    cv2.line(temp_img, self.polygon_vertices[i], self.polygon_vertices[i + 1], self.clr, self.thickness)
                 cv2.imshow(self.title, temp_img)
         elif event == cv2.EVENT_LBUTTONUP:
             self.drawing = False
             temp_img = self.image.copy()
             for i in range(len(self.polygon_vertices) - 1):
-                cv2.line(
-                    temp_img,
-                    self.polygon_vertices[i],
-                    self.polygon_vertices[i + 1],
-                    self.clr,
-                    self.thickness,
-                )
+                cv2.line(temp_img, self.polygon_vertices[i], self.polygon_vertices[i + 1], self.clr, self.thickness)
             cv2.imshow(self.title, temp_img)
 
     def run(self):
@@ -112,37 +101,32 @@ class ROISelectorPolygon(object):
         cv2.setMouseCallback(self.title, self.draw_polygon)
         while True:
             img_copy = self.image.copy()
+            for i in range(len(self.polygon_vertices)):
+                cv2.circle(img_copy, center=self.polygon_vertices[i], radius=self.vertice_size, color=self.clr, thickness=-1)
             for i in range(len(self.polygon_vertices) - 1):
-                cv2.line(
-                    img_copy,
-                    self.polygon_vertices[i],
-                    self.polygon_vertices[i + 1],
-                    self.clr,
-                    self.thickness,
-                )
+                cv2.line(img_copy, self.polygon_vertices[i], self.polygon_vertices[i + 1], self.clr, self.thickness)
             if self.drawing and len(self.polygon_vertices) > 0:
-                cv2.line(
-                    img_copy,
-                    self.polygon_vertices[-1],
-                    (self.polygon_vertices[-1][0], self.polygon_vertices[-1][1]),
-                    self.clr,
-                    self.thickness,
-                )
+                cv2.line(img_copy, self.polygon_vertices[-1], (self.polygon_vertices[-1][0], self.polygon_vertices[-1][1]), self.clr, self.thickness,)
             cv2.imshow(self.title, img_copy)
             key = cv2.waitKey(1)
             if key in [27, ord("q"), ord("Q"), ord(" ")]:
                 if self.run_checks():
-                    cv2.destroyAllWindows()
+                    if self.destroy:
+                        cv2.destroyAllWindows()
                     cv2.waitKey(1)
-                    self.polygon_vertices = np.array(self.polygon_vertices)
+                    self.polygon = Polygon(np.array(self.polygon_vertices)).simplify(tolerance=20, preserve_topology=True)
+                    self.polygon_vertices = np.array(self.polygon.exterior.coords)
+                    self.polygon_area = self.polygon.area
+                    self.polygon_arr = np.array(self.polygon.exterior.coords).astype(np.int32)[1:]
+                    self.max_vertice_distance = np.max(cdist(self.polygon_vertices, self.polygon_vertices).astype(np.int32))
+                    self.polygon_centroid = np.array(self.polygon.centroid).astype(int)
+                    self.tags = {f'Tag_{cnt}': tuple(y) for cnt, y in enumerate(self.polygon_arr)}
+                    self.tags['Center_tag'] = tuple(self.polygon_centroid)
                     break
 
     def run_checks(self):
         if len(list(set(self.polygon_vertices))) < 3:
-            CropWarning(
-                msg="CROP WARNING: At least 3 unique vertices are needed to form a polygon. Please try again.",
-                source=self.__class__.__name__,
-            )
+            ROIWarning(msg="ROI WARNING: At least 3 unique vertices are needed to form a polygon. Please try again.", source=self.__class__.__name__)
             return False
         else:
             return True
